Move update task run model method
